Share the axe scan itself, not just the failure formatter
docs had scanArticle while www re-implemented the AxeBuilder chain inline, so
the two suites configured axe differently for no reason. Add scan() to
e2e/shared/axe.ts, taking rules, tags, excludeRules, and include, and have both
suites call it. settleForAxe moves across too, and axe-helpers.ts drops to the
parts that are genuinely docs-specific: its rule lists, the article-scoped
result shape, the warn-mode split, and report attachment.

Rules stay per suite. docs enforces heading-order and page-has-heading-one; www
enforces only page-has-heading-one. That reverses the previous commit, which
enforced heading-order on www: a full-site scan of all 480 content pages found
it violated on 437 of them, nearly all from the footer's h6 column headings and
the related-posts card's h4. Enforcing it now would fail almost every pull
request, so it waits on those two components being fixed.
