+++ e2e/shared/package.json
@@ -3,6 +3,11 @@
   "version": "1.0.0",
   "private": true,
   "type": "module",
+  "dependencies": {
+    "@axe-core/playwright": "^4.12.1",
+    "@playwright/test": "^1.59.1",
+    "axe-core": "^4.12.1"
+  },
   "devDependencies": {
     "@types/node": "catalog:"
   }
